Replace non-optimal HIP porting (__syncthreads for __syncwarp)
     with a more performant HIP porting (using __threadfence_block).

     On AMD Instinct MI-250 system (optimization in progress),
     this increases performance by ~10% in case fp32 weights.