@HaiShaw 2022-09-15
@sryap Thanks for your asking. 

Have UTs and Performance comparison data below:



**These related UTs pass:**

/workspace/FBGEMM/fbgemm_gpu# python test/split_table_batched_embeddings_test.py SplitTableBatchedEmbeddingsTest.test_nbit_forward_fused_pooled_emb_quant
/workspace/FBGEMM/fbgemm_gpu# python test/split_table_batched_embeddings_test.py SplitTableBatchedEmbeddingsTest.test_forward_fused_pooled_emb_quant

**Performance results**

**without this change**
$:/workspace/FBGEMM/fbgemm_gpu# python /workspace/FBGEMM/fbgemm_gpu/bench/split_table_batched_embeddings_benchmark.py nbit-device --num-tables=2 --embedding-dim=192 --num-embeddings=10000000 --batch-size=65536 --iters=100 --row-wise --alpha=1.0 --bag-size=32 --weights-precision=fp32 --output-dtype=int8
INFO:root:fp32 Embedding tables: 20000000 rows, 3.84 GParam, 15.36 GB
INFO:root:Accessed weights per batch: 4194304 rows, 3.22 GB
INFO:root:Iteration 0: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 568.95 GB/s, Time: 5706us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 1: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 568.85 GB/s, Time: 5707us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 2: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 570.49 GB/s, Time: 5691us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 3: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 568.87 GB/s, Time: 5707us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 4: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 568.77 GB/s, Time: 5708us, Memory Usage For Pruning: 0 GB
INFO:root:Average of all iterations: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 569.37 GB/s, Time: 5702us, Memory Usage For Pruning: 0 GB
$:/workspace/FBGEMM/fbgemm_gpu#
$:/workspace/FBGEMM/fbgemm_gpu#
$:/workspace/FBGEMM/fbgemm_gpu# python /workspace/FBGEMM/fbgemm_gpu/bench/split_table_batched_embeddings_benchmark.py nbit-device --num-tables=2 --embedding-dim=192 --num-embeddings=10000000 --batch-size=65536 --iters=100 --row-wise --alpha=1.0 --bag-size=32 --weights-precision=fp32 --output-dtype=fp16
INFO:root:fp32 Embedding tables: 20000000 rows, 3.84 GParam, 15.36 GB
INFO:root:Accessed weights per batch: 4194304 rows, 3.22 GB
INFO:root:Iteration 0: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 600.42 GB/s, Time: 5449us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 1: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 601.96 GB/s, Time: 5435us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 2: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 602.02 GB/s, Time: 5434us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 3: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 600.24 GB/s, Time: 5450us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 4: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 600.19 GB/s, Time: 5451us, Memory Usage For Pruning: 0 GB
INFO:root:Average of all iterations: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 600.82 GB/s, Time: 5445us, Memory Usage For Pruning: 0 GB
$:/workspace/FBGEMM/fbgemm_gpu#
$:/workspace/FBGEMM/fbgemm_gpu#
$:/workspace/FBGEMM/fbgemm_gpu# python /workspace/FBGEMM/fbgemm_gpu/bench/split_table_batched_embeddings_benchmark.py nbit-device --num-tables=2 --embedding-dim=192 --num-embeddings=10000000 --batch-size=65536 --iters=100 --row-wise --alpha=1.0 --bag-size=32 --weights-precision=fp32 --output-dtype=fp32
INFO:root:fp32 Embedding tables: 20000000 rows, 3.84 GParam, 15.36 GB
INFO:root:Accessed weights per batch: 4194304 rows, 3.22 GB
INFO:root:Iteration 0: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 609.31 GB/s, Time: 5452us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 1: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 611.00 GB/s, Time: 5437us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 2: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 610.90 GB/s, Time: 5438us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 3: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 610.86 GB/s, Time: 5438us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 4: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 610.86 GB/s, Time: 5438us, Memory Usage For Pruning: 0 GB
INFO:root:Average of all iterations: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 610.87 GB/s, Time: 5438us, Memory Usage For Pruning: 0 GB

**With this change**
$:/workspace/FBGEMM/fbgemm_gpu# python /workspace/FBGEMM/fbgemm_gpu/bench/split_table_batched_embeddings_benchmark.py nbit-device --num-tables=2 --embedding-dim=192 --num-embeddings=10000000 --batch-size=65536 --iters=100 --row-wise --alpha=1.0 --bag-size=32 --weights-precision=fp32 --output-dtype=int8
INFO:root:fp32 Embedding tables: 20000000 rows, 3.84 GParam, 15.36 GB
INFO:root:Accessed weights per batch: 4194304 rows, 3.22 GB
INFO:root:Iteration 0: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 620.25 GB/s, Time: 5234us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 1: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 619.92 GB/s, Time: 5237us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 2: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 619.77 GB/s, Time: 5238us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 3: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 622.00 GB/s, Time: 5219us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 4: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 622.05 GB/s, Time: 5219us, Memory Usage For Pruning: 0 GB
INFO:root:Average of all iterations: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 621.27 GB/s, Time: 5225us, Memory Usage For Pruning: 0 GB
$:/workspace/FBGEMM/fbgemm_gpu#
$:/workspace/FBGEMM/fbgemm_gpu#
$:/workspace/FBGEMM/fbgemm_gpu# python /workspace/FBGEMM/fbgemm_gpu/bench/split_table_batched_embeddings_benchmark.py nbit-device --num-tables=2 --embedding-dim=192 --num-embeddings=10000000 --batch-size=65536 --iters=100 --row-wise --alpha=1.0 --bag-size=32 --weights-precision=fp32 --output-dtype=fp16
INFO:root:fp32 Embedding tables: 20000000 rows, 3.84 GParam, 15.36 GB
INFO:root:Accessed weights per batch: 4194304 rows, 3.22 GB
INFO:root:Iteration 0: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 654.83 GB/s, Time: 4996us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 1: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 655.06 GB/s, Time: 4994us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 2: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 657.25 GB/s, Time: 4978us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 3: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 657.41 GB/s, Time: 4976us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 4: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 657.09 GB/s, Time: 4979us, Memory Usage For Pruning: 0 GB
INFO:root:Average of all iterations: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 657.25 GB/s, Time: 4978us, Memory Usage For Pruning: 0 GB
$:/workspace/FBGEMM/fbgemm_gpu#
$:/workspace/FBGEMM/fbgemm_gpu#
$:/workspace/FBGEMM/fbgemm_gpu# python /workspace/FBGEMM/fbgemm_gpu/bench/split_table_batched_embeddings_benchmark.py nbit-device --num-tables=2 --embedding-dim=192 --num-embeddings=10000000 --batch-size=65536 --iters=100 --row-wise --alpha=1.0 --bag-size=32 --weights-precision=fp32 --output-dtype=fp32
INFO:root:fp32 Embedding tables: 20000000 rows, 3.84 GParam, 15.36 GB
INFO:root:Accessed weights per batch: 4194304 rows, 3.22 GB
INFO:root:Iteration 0: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 666.01 GB/s, Time: 4988us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 1: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 665.76 GB/s, Time: 4990us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 2: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 665.64 GB/s, Time: 4991us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 3: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 665.76 GB/s, Time: 4990us, Memory Usage For Pruning: 0 GB
INFO:root:Iteration 4: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 665.67 GB/s, Time: 4990us, Memory Usage For Pruning: 0 GB
INFO:root:Average of all iterations: fp32 Forward, B: 65536, E: 10000000, T: 2, D: 192, L: 32, W: False, BW: 665.69 GB/s, Time: 4990us, Memory Usage For Pruning: 0 GB